// Microsoft.MyWorkspace.UI/src/types/AzureWorkspace/VMPatchDetailsDto.types.ts
export interface VMPatchDetails {
  Id: string;
  ResourceGroup: string;
  VirtualMachineId: string;
  WorkspaceId: string;
  Title: string;
  KBID: string;
  Product: string;
  MSRCSeverity: string;
  PublishedDate: string;
  SourceSystem: string;
  TimeGenerated: string;
  Classification: string;
}